# server/openapi_spec.yaml
openapi: 3.0.0
info:
  title: Astronova Backend API
  description: Comprehensive astrology API with AI-powered insights, birth charts, compatibility matching, and personalized reports
  version: 1.0.0
  contact:
    name: Astronova Support
    email: admin@100xai.engineering
servers:
  - url: https://astronova-ghcr.onrender.com
    description: Production server
  - url: http://localhost:8080
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: "Admin API token. Production admin endpoints also accept Authorization: Bearer <ADMIN_API_TOKEN>."
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        appleUserId:
          type: string
...
paths:
...
  /api/v1/auth/apple:
    post:
      summary: Authenticate with Apple ID
      description: Requires an Apple-issued idToken or identityToken. Guest/anonymous mode must not call this endpoint.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              oneOf:
                - required:
                    - idToken
                - required:
                    - identityToken
              properties:
                idToken:
                  type: string
                  description: Apple Sign-In identityToken sent by current iOS clients.
                identityToken:
                  type: string
                  description: Alias for idToken.
                userIdentifier:
                  type: string
                  description: Client-observed Apple user identifier. The validated token subject is authoritative.
                email:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwtToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  expiresAt:
                    type: string
                    format: date-time
        '401':
          description: Missing or invalid Apple ID token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/validate:
    get:
      summary: Validate JWT token
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
...
  /api/v1/auth/refresh:
    post:
      summary: Refresh JWT token
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Refreshed token
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwtToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  expiresAt:
                    type: string
                    format: date-time
        '401':
          description: Authorization required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...